Item products

+++ database/migrations/2023_04_09_075721_create_items_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->foreignId('Category_id');
             $table->string('name');
             $table->text('description');
-            $table->integer('price');
+            $table->decimal('price', 16, 2);
             $table->string('image');
             $table->timestamps();
         });
